make defining cli enums a macro
